requirements.txt
rich>=13.7.0
bcrypt>=4.1.2
argon2-cffi>=23.1.0
